backend/requirements.txt
Django==4.1.3
base58
djangorestframework
psycopg2-binary 
python-dotenv
django-cors-headers
django_redis
requests
eth-account
